Compliace with ecj, checkstyle and formatter
