+++ truffle/com.oracle.truffle.api.boot/src/com/oracle/truffle/api/boot/TruffleInfo.java
@@ -1,7 +1,8 @@
 package com.oracle.truffle.api.boot;
 
-/** Provides informations about Truffle runtime to
- * implementation of {@link TruffleServices}. Obtain via {@link TruffleServices#info()}.
+/**
+ * Provides informations about Truffle runtime to implementation of {@link TruffleServices}. Obtain
+ * via {@link TruffleServices#info()}.
  */
 public abstract class TruffleInfo {
     static TruffleInfo DEFAULT;
@@ -13,16 +14,27 @@ public abstract class TruffleInfo {
         }
     }
 
-    /** Restricted constructor. Available only to selected subclasses.
+    /**
+     * Restricted constructor. Available only to selected subclasses.
+     * 
      * @throws IllegalStateException usually throws an exception
      */
     protected TruffleInfo() {
-        if (DEFAULT != null) {
+        if (DEFAULT != null || !getClass().getName().equals("com.oracle.truffle.api.impl.TruffleInfoImpl")) {
             throw new IllegalStateException();
         }
         DEFAULT = this;
     }
 
+    /**
+     * Allows the Truffle API to access services of the runtime. For example
+     * {@link LoopCountSupport}, etc.
+     *
+     * @param <T> the requested type
+     * @param type class of the request service
+     * @return instance of the requested type or <code>null</code> if none has been found
+     */
+    @SuppressWarnings({"unchecked", "static-method"})
     protected final <T> T lookup(Class<T> type) {
         if (type == LoopCountSupport.class) {
             return (T) TruffleServices.DEFAULT.loopCount();
@@ -31,5 +43,6 @@ public abstract class TruffleInfo {
     }
 
     public abstract Class<?> findLanguage(Object node);
+
     public abstract void initializeCallTarget(Object callTarget);
 }
